refactor(api): v0 consolidation
