+++ packages/app/src/pages/project/[project_id]/tasks/update.astro
@@ -1,5 +1,5 @@
 ---
-import { getProject, getRecentUpdate, getProjectTasks, type Task } from "@devpad/core";
+import { getProject, getRecentUpdate, getProjectTasks, type Task, _FetchedTask, TaskWithDetails } from "@devpad/core";
 import PageLayout from '@/layouts/PageLayout.astro';
 import ProjectTasksLayout from '@/layouts/ProjectTasksLayout.astro';
 import { UpdateDiffList } from '@/components/solid/UpdateDiff';
@@ -23,7 +23,7 @@ const codebase_task_map = project_tasks.reduce(
 		acc[task.codebase_tasks.id] = task;
 		return acc;
 	},
-	{} as Record<string, Task>
+	{} as Record<string, TaskWithDetails>
 );
 
 // get the most recent todo_update for the project and render it
@@ -35,7 +35,7 @@ const update_data = update?.data ? (typeof update.data === "string" ? JSON.parse
 <PageLayout title={`Update Codebase - ${project_id}`}>
 	<ProjectTasksLayout project_id={project_id}>
 		<div class="flex-col" style="gap: 10px">
-			<a role="button" id="update-check"> check for updates </a>
+			<a role="button" id="update-check" data-project-id={project_uuid}> check for updates </a>
 			<p id="output"></p>
 			<div></div>
 			{update ? <UpdateDiffList items={update_data ?? []} tasks={codebase_task_map} project_id={project_uuid} update_id={update.id} client:load /> : <p>No updates</p>}
@@ -44,14 +44,17 @@ const update_data = update?.data ? (typeof update.data === "string" ? JSON.parse
 </PageLayout>
 
 <script is:inline define:vars={{ project_uuid }}>
+	import { getApiClient } from "../../../../utils/api-client";
+	
 	function hook_check() {
 		const update_check = document.getElementById("update-check");
 		if (!update_check) return;
 		const output = document.getElementById("output");
 		if (!output) throw new Error("Output element not found");
 		update_check.addEventListener("click", async () => {
-			const response = await fetch(`/api/project/scan?project_id=${project_uuid}`, { method: "POST" });
-			if (!response.ok) {
+			const apiClient = getApiClient();
+			const response = await apiClient.projects.scan.start(project_uuid);
+			if (!response.ok || !response.body) {
 				output.innerText = "Error: " + response.statusText;
 				return;
 			}
